pyproject.toml
[project]
name = "langchain-ex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.45.0",
    "langchain>=0.3.27",
    "langchain-ollama>=0.2.1",
    "langchain-openai>=0.2.1",
    "langchain-anthropic>=0.2.1",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]
